% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_surf.R
\name{as.cimg.magick-image}
\alias{as.cimg.magick-image}
\title{Convert magick-image to cimg}
\usage{
"as.cimg.magick-image"(im)
}
\arguments{
\item{im}{an object of class magick-image from the magick package}
}
\description{
Convert magick-image to cimg. Put in this package as long as it is not in the imager package
}
