% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{image.dlib-package}
\alias{image.dlib-package}
\title{Image Processing functionality using the 'dlib' package.}
\description{
Facility wrappers around the image processing functionality of 'dlib'. 'Dlib' <http://dlib.net> is a 'C++' toolkit containing machine learning algorithms and computer vision tools. Currently the package allows to find SURF descriptors.
}
\seealso{
\code{\link{image_surf}}, \code{\link{image_fhog}}
}
