% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_fhog.R
\name{image_fhog}
\alias{image_fhog}
\title{HOG (Histogram Of Gradients) feature extraction}
\usage{
image_fhog(file, cell_size = 8L, filter_rows_padding = 1L,
  filter_cols_padding = 1L)
}
\arguments{
\item{file}{a bmp file in bmp3 format with extension .bmp}

\item{cell_size}{integer. HOG groups pixels in cells, based on which a histogram of the gradiens is computed. Give the cell size. Defaults to 8.}

\item{filter_rows_padding}{integer indicating how many extra rows and columns of zero padding along the borders to add, for more efficient convolution code}

\item{filter_cols_padding}{integer indicating how many extra rows and columns of zero padding along the borders to add, for more efficient convolution code}
}
\value{
a list with HOG features of the pixels mapped into HOG space. It contains:
\itemize{
 \item{hog_height: }{The size of the height of the image in HOG space}
 \item{hog_width: }{The size of the width of the image in HOG space}
 \item{hog_feature: }{The 31 dimensional FHOG vector, one for each pixel in HOG space. This is an array of dimension hog_height x hog_width x 31. The 31-dimensional vector describes the gradient structure within the cell.}
 \item{hog_cell_size: }{The HOG cell_size}
 \item{filter_rows_padding: }{The value of filter_rows_padding}
 \item{filter_cols_padding: }{The value of filter_cols_padding}
}
}
\description{
Gets the histogram of oriented gradients (HOG) feature descriptor
\url{https://en.wikipedia.org/wiki/Histogram_of_oriented_gradients} of an image.
More information on the calculation, see the reference paper or
the dlib reference \url{http://dlib.net/dlib/image_transforms/fhog_abstract.h.html}
}
\examples{
f <- system.file("extdata", "cruise_boat.bmp", package="image.dlib")
x <- image_fhog(f, cell_size = 8)
str(x)

## FHOG feature vector of pixel at HOG space position 1, 1
x$fhog[1, 1, ]
## FHOG feature vector of pixel at HOG space position 4 (height), 7 (width)
x$fhog[4, 7, ]
}
\references{
Object Detection with Discriminatively Trained Part Based Models by P. Felzenszwalb, R. Girshick, D. McAllester, D. Ramanan IEEE Transactions on Pattern Analysis and Machine Intelligence, Vol. 32, No. 9, Sep. 2010
}
